asr-leaderboard/eval/requirements.txt
# No external deps required (pure Python).
PyYAML>=6.0
opencc-python-reimplemented>=0.1.7